Using native JSON operations to add thread participants (#18093)
Automatic Merge
